+++ src/plugins/widget.js
@@ -77,6 +77,7 @@ export default class PluginWidget {
     this.editableView = options.editableView
     this.pm = pm
     this.widgetPath = options.widgetPath
+    this.coverPrefs = options.coverPrefs || {}
 
     this.widgets = {}
     this.el = document.createElement('div')
@@ -228,6 +229,7 @@ export default class PluginWidget {
       , initialRectangle: rectangle
       , initialFocus
       , widgetPath: this.widgetPath
+      , coverPrefs: this.coverPrefs
       }
     )